Revert "fix: SMALLINT 컬럼(age/weekStartDay/priority) 매핑을 Integer→Short로 변경"
This reverts commit 9143e5c64e879ed2ac92b40fdc7b8efb59d60f8d.

+++ backend/src/main/java/com/ssafy/s309/domain/user/dto/SettingsResponse.java
@@ -6,7 +6,7 @@ import com.ssafy.s309.domain.user.entity.User;
 public record SettingsResponse(
     Long userId,
     String name,
-    Short age,
+    Integer age,
     String gender,
     String phone,
     Float height,
@@ -15,7 +15,7 @@ public record SettingsResponse(
     Boolean isMedicated,
     Integer targetLow,
     Integer targetHigh,
-    Short weekStartDay) {
+    Integer weekStartDay) {
 
   public static SettingsResponse from(User user) {
     return new SettingsResponse(
